pyproject.toml
[tool.poetry]
name = "hellobot"
version = "0.1.0"
description = ""
authors = ["Чичко Александра Алексеевна <aachichko@sberautotech.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^5.4.1"
slackeventsapi = "^3.0.0"
bs4 = "^0.0.1"
pydantic = "^1.8.2"
Flask = "^2.0.2"
slack = "^0.0.2"

[tool.poetry.dev-dependencies]
isort = "^5.9.3"
black = "^21.7b0"
pre-commit = "^2.14.0"
pytest = "^6.2.4"

[tool.black]
line_length = 120
multi_line_output = 3
length_sort = true
target-version = ["py38"]
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -s -vv"
testpaths = [
    "tests",
    "test_*",
    "tests_*",
    "*_test",
    "*_tests"
]
python_files = [
    "tests.py",
    "test_*.py",
    "tests_*.py",
    "*_test.py",
    "*_tests.py"
]

[tool.poetry.scripts]
vr = "hellobot.main:main"

[build-system]
requires = ["poetry-core^1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry.core.masonry.api"